+++ device_config.py
@@ -1,0 +1,2 @@
+camera_size = (1280, 720)
+screen_size = (1920, 1080)
